requirements.txt
python-telegram-bot==20.3
pycryptodome==3.19.0
aiosqlite==0.19.0
